backend/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.16
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
google-generativeai==0.8.3
